Project/Rev_1/requirements.txt
altair
matplotlib
streamlit
pandas
seaborn == 0.12.2
countryinfo
plotly.express
